gateway: migrate container wrappers
File size update:

```
    7512 src/omero/gateway/__init__.py
     436 src/omero/gateway/_annotations.py
     754 src/omero/gateway/_containers.py
    1623 src/omero/gateway/_core.py
     619 src/omero/gateway/_instruments.py
       1 src/omero/gateway/_wrappers.py
   10828 src/omero/gateway/old.py
```
